fixup! Add writer
